Refactoring upload doc form into modal
