+++ src/routes/Documents.jsx
@@ -1,14 +1,15 @@
 // React Imports
-import React, { useContext } from 'react';
+import React, { useContext, useState } from 'react';
 import { useLocation } from 'react-router-dom';
 // Material UI Imports
+import AddIcon from '@mui/icons-material/Add';
+import Box from '@mui/material/Box';
 import Button from '@mui/material/Button';
 import Remove from '@mui/icons-material/Remove';
-import Box from '@mui/material/Box';
 import Typography from '@mui/material/Typography';
 // Component Imports
 import {
-  UploadDocumentForm,
+  UploadDocumentModal,
   SetAclPermissionForm,
   SetAclPermsDocContainerForm
 } from '../components/Form';
@@ -23,6 +24,8 @@ import DocumentTable from '../components/Documents/DocumentTable';
  * @returns {React.JSX.Element} The Documents Page
  */
 const Documents = () => {
+  // state for UploadDocumentModal component
+  const [showModal, setShowModal] = useState(false);
   const location = useLocation();
 
   localStorage.setItem('restorePath', location.pathname);
@@ -57,7 +60,21 @@ const Documents = () => {
         <Typography>Client selected: {selectedUser.person || selectedUser.podUrl}</Typography>
       )}
 
-      <UploadDocumentForm />
+      <Button
+        variant="contained"
+        color="secondary"
+        size="small"
+        aria-label="Add Client Button"
+        startIcon={<AddIcon />}
+        onClick={() => setShowModal(true)}
+        sx={{ marginTop: '3rem' }}
+      >
+        Add Client
+      </Button>
+      {/* modal/popup renders when setShowModal state is true */}
+      <UploadDocumentModal showModal={showModal} setShowModal={setShowModal} />
+
+      {/* <UploadDocumentModal /> */}
       <DocumentTable />
       <SetAclPermsDocContainerForm />
       <SetAclPermissionForm />
